DUOS-919 - Remove ConsentAPI and move to ConsentService
