+++ src/test/java/org/broadinstitute/consent/http/resources/DataAccessRequestResourceTest.java
@@ -19,11 +19,10 @@ import org.broadinstitute.consent.http.models.DataAccessRequestData;
 import org.broadinstitute.consent.http.models.DataAccessRequestManage;
 import org.broadinstitute.consent.http.models.DataSet;
 import org.broadinstitute.consent.http.models.User;
-import org.broadinstitute.consent.http.service.AbstractConsentAPI;
 import org.broadinstitute.consent.http.service.AbstractDataAccessRequestAPI;
 import org.broadinstitute.consent.http.service.AbstractElectionAPI;
 import org.broadinstitute.consent.http.service.AbstractMatchProcessAPI;
-import org.broadinstitute.consent.http.service.ConsentAPI;
+import org.broadinstitute.consent.http.service.ConsentService;
 import org.broadinstitute.consent.http.service.DataAccessRequestAPI;
 import org.broadinstitute.consent.http.service.DataAccessRequestService;
 import org.broadinstitute.consent.http.service.ElectionAPI;
@@ -46,7 +45,6 @@ import org.powermock.modules.junit4.PowerMockRunner;
 @PowerMockIgnore("jdk.internal.reflect.*")
 @PrepareForTest({
         AbstractDataAccessRequestAPI.class,
-        AbstractConsentAPI.class,
         AbstractMatchProcessAPI.class,
         AbstractUseRestrictionValidatorAPI.class,
         AbstractDACUserAPI.class,
@@ -63,7 +61,7 @@ public class DataAccessRequestResourceTest {
     @Mock
     private EmailNotifierService emailNotifierService;
     @Mock
-    private ConsentAPI consentAPI;
+    private ConsentService consentService;
     @Mock
     private DataAccessRequestAPI dataAccessRequestAPI;
     @Mock
@@ -79,7 +77,6 @@ public class DataAccessRequestResourceTest {
     public void setUp() {
         MockitoAnnotations.initMocks(this);
         PowerMockito.mockStatic(AbstractDataAccessRequestAPI.class);
-        PowerMockito.mockStatic(AbstractConsentAPI.class);
         PowerMockito.mockStatic(AbstractMatchProcessAPI.class);
         PowerMockito.mockStatic(AbstractUseRestrictionValidatorAPI.class);
         PowerMockito.mockStatic(AbstractDACUserAPI.class);
@@ -93,14 +90,13 @@ public class DataAccessRequestResourceTest {
     public void testDescribeConsentForDarCase1() {
         DataAccessRequest dar = generateDataAccessRequest();
         when(dataAccessRequestService.findByReferenceId(any())).thenReturn(dar);
-        when(consentAPI.getConsentFromDatasetID(any())).thenReturn(new Consent());
+        when(consentService.getConsentFromDatasetID(any())).thenReturn(new Consent());
         when(AbstractDataAccessRequestAPI.getInstance()).thenReturn(dataAccessRequestAPI);
-        when(AbstractConsentAPI.getInstance()).thenReturn(consentAPI);
         when(AbstractDACUserAPI.getInstance()).thenReturn(dacUserAPI);
         when(AbstractElectionAPI.getInstance()).thenReturn(electionAPI);
         when(user.getDacUserId()).thenReturn(dar.getUserId());
         when(userService.findUserByEmail(any())).thenReturn(user);
-        resource = new DataAccessRequestResource(dataAccessRequestService, emailNotifierService, userService);
+        resource = new DataAccessRequestResource(dataAccessRequestService, emailNotifierService, userService, consentService);
         Consent consent = resource.describeConsentForDAR(authUser, dar.getReferenceId());
         assertNotNull(consent);
     }
@@ -114,14 +110,13 @@ public class DataAccessRequestResourceTest {
         when(dataAccessRequestService.findByReferenceId(any())).thenReturn(dar);
         DataSet dataSet = new DataSet();
         dataSet.setDataSetId(1);
-        when(consentAPI.getConsentFromDatasetID(any())).thenReturn(new Consent());
+        when(consentService.getConsentFromDatasetID(any())).thenReturn(new Consent());
         when(AbstractDataAccessRequestAPI.getInstance()).thenReturn(dataAccessRequestAPI);
-        when(AbstractConsentAPI.getInstance()).thenReturn(consentAPI);
         when(AbstractDACUserAPI.getInstance()).thenReturn(dacUserAPI);
         when(AbstractElectionAPI.getInstance()).thenReturn(electionAPI);
         when(user.getDacUserId()).thenReturn(dar.getUserId());
         when(userService.findUserByEmail(any())).thenReturn(user);
-        resource = new DataAccessRequestResource(dataAccessRequestService, emailNotifierService, userService);
+        resource = new DataAccessRequestResource(dataAccessRequestService, emailNotifierService, userService, consentService);
         Consent consent = resource.describeConsentForDAR(authUser, dar.getReferenceId());
         assertNotNull(consent);
     }
@@ -133,12 +128,11 @@ public class DataAccessRequestResourceTest {
     public void testDescribeConsentForDarCase3() {
         DataAccessRequest dar = generateDataAccessRequest();
         when(dataAccessRequestService.findByReferenceId(any())).thenReturn(dar);
-        when(consentAPI.getConsentFromDatasetID(any())).thenReturn(null);
+        when(consentService.getConsentFromDatasetID(any())).thenReturn(null);
         when(AbstractDataAccessRequestAPI.getInstance()).thenReturn(dataAccessRequestAPI);
-        when(AbstractConsentAPI.getInstance()).thenReturn(consentAPI);
         when(AbstractDACUserAPI.getInstance()).thenReturn(dacUserAPI);
         when(AbstractElectionAPI.getInstance()).thenReturn(electionAPI);
-        resource = new DataAccessRequestResource(dataAccessRequestService, emailNotifierService, userService);
+        resource = new DataAccessRequestResource(dataAccessRequestService, emailNotifierService, userService, consentService);
         resource.describeConsentForDAR(authUser, dar.getReferenceId());
     }
 
@@ -153,7 +147,7 @@ public class DataAccessRequestResourceTest {
         when(AbstractDataAccessRequestAPI.getInstance()).thenReturn(dataAccessRequestAPI);
         when(AbstractDACUserAPI.getInstance()).thenReturn(dacUserAPI);
         when(AbstractElectionAPI.getInstance()).thenReturn(electionAPI);
-        resource = new DataAccessRequestResource(dataAccessRequestService, emailNotifierService, userService);
+        resource = new DataAccessRequestResource(dataAccessRequestService, emailNotifierService, userService, consentService);
         resource.describeConsentForDAR(authUser, dar.getReferenceId());
     }
 
@@ -166,14 +160,13 @@ public class DataAccessRequestResourceTest {
         when(dataAccessRequestService.findByReferenceId(any())).thenReturn(dar);
         DataSet dataSet = new DataSet();
         dataSet.setDataSetId(1);
-        when(consentAPI.getConsentFromDatasetID(any())).thenReturn(new Consent());
+        when(consentService.getConsentFromDatasetID(any())).thenReturn(new Consent());
         when(AbstractDataAccessRequestAPI.getInstance()).thenReturn(dataAccessRequestAPI);
-        when(AbstractConsentAPI.getInstance()).thenReturn(consentAPI);
         when(AbstractDACUserAPI.getInstance()).thenReturn(dacUserAPI);
         when(AbstractElectionAPI.getInstance()).thenReturn(electionAPI);
         when(user.getDacUserId()).thenReturn(dar.getUserId() + 1);
         when(userService.findUserByEmail(any())).thenReturn(user);
-        resource = new DataAccessRequestResource(dataAccessRequestService, emailNotifierService, userService);
+        resource = new DataAccessRequestResource(dataAccessRequestService, emailNotifierService, userService, consentService);
         resource.describeConsentForDAR(authUser, dar.getReferenceId());
     }
 
@@ -187,7 +180,8 @@ public class DataAccessRequestResourceTest {
         resource = new DataAccessRequestResource(
             dataAccessRequestService,
             emailNotifierService,
-            userService);
+            userService,
+            consentService);
         Response response = resource.describeManageDataAccessRequestsV2(authUser);
         assertEquals(HttpStatusCodes.STATUS_CODE_OK, response.getStatus());
     }
